+++ change-log.md
@@ -1,10 +1,12 @@
 ## Check Points
 
+### Auth
 - next installation 
 - shadcn installation
 - shadcn component installation
 - shadcn theme installation, font setup
 
+### Database + Auth
 - clerk auth 
     > might change to custom auth
 - convex database
@@ -16,6 +18,16 @@
     > schema philosophy https://docs.convex.dev/database/advanced/schema-philosophy
 - how we actually connected convex & clerk
 - Auth pages
+
+### Background Jobs
+
+- setup ai sdk 
+- create blocking api route
+- setup inngest sdk
+- create backgronud job 
+- blocking vs non-blocking
+
+
 
 
 ---